requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-filter==1.1.0
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-filters==0.11.1
filter==0.0.0.20200724
Pillow==9.0.1
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
